refactor: Add src/shared
