+++ src/frontend/plugins/edusharing-asset.tsx
@@ -10,7 +10,7 @@ import Modal from 'react-modal'
 import Image from 'next/image'
 import { MouseEvent } from 'react'
 import { useEffect, useRef, useState } from 'react'
-import { EdusharingAssetDecoder } from '../../decoders'
+import { EdusharingAssetDecoder } from '../../shared/decoders'
 
 const state = object({
   edusharingAsset: optional(
