+++ src/pages/index.ts
@@ -1,7 +1,7 @@
 import { GetServerSideProps } from 'next'
 
-import { kitchenSink } from '../storage-format/kitchen-sink'
-import { migrate, emptyDocument } from '../storage-format'
+import { kitchenSink } from '../shared/storage-format/kitchen-sink'
+import { migrate, emptyDocument } from '../shared/storage-format'
 import { SerloEditor, SerloEditorProps } from '../frontend'
 
 export const getServerSideProps: GetServerSideProps<
