+++ src/shared/storage-format/kitchen-sink.ts
