+++ tsconfig.server.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["src/backend/*.ts", "src/server-utils.ts", "src/decoders.ts"],
+  "include": ["src/backend/*.ts", "src/server-utils.ts", "src/shared/**/*.ts"],
   "compilerOptions": {
     "allowJs": false,
     "noEmit": false
